Agrego nueva navegabilidad al nuevo icono servicios y a los botones dentro de dicha activity
