+++ app/src/main/java/com/dpozzo68/analizarapp/ContactoActivity.java
@@ -64,13 +64,15 @@ public class ContactoActivity extends AppCompatActivity {
         imagen.setClickable(true);
         Intent intent = new Intent(this, MisConsumos.class);
         startActivity(intent);
+        finish();
     }
 
-    public void irAlarmas(View view) {
-        ImageView imagen = findViewById(R.id.imagen_alerta);
+    public void irServicios(View view) {
+        ImageView imagen = findViewById(R.id.imagen_servicios);
         imagen.setClickable(true);
-        Intent intent = new Intent(this, MisAlarmas1.class);
+        Intent intent = new Intent(this, Servicios.class);
         startActivity(intent);
+        finish();
     }
 
     public void irPerfil(View view) {
@@ -78,6 +80,7 @@ public class ContactoActivity extends AppCompatActivity {
         imagen.setClickable(true);
         Intent intent = new Intent(this, Mi_Cuenta.class);
         startActivity(intent);
+        finish();
     }
 
     public void irContacto(View view) {
@@ -85,6 +88,7 @@ public class ContactoActivity extends AppCompatActivity {
         imagen.setClickable(true);
         Intent intent = new Intent(this, ContactoActivity.class);
         startActivity(intent);
+        finish();
     }
 
     @Override
