+++ app/src/main/java/com/dpozzo68/analizarapp/Mi_Cuenta.java
@@ -103,13 +103,15 @@ public class Mi_Cuenta extends AppCompatActivity {
         imagen.setClickable(true);
         Intent intent = new Intent(this, MisConsumos.class);
         startActivity(intent);
+        finish();
     }
 
-    public void irAlarmas(View view) {
-        ImageView imagen = findViewById(R.id.imagen_alerta);
+    public void irServicios(View view) {
+        ImageView imagen = findViewById(R.id.imagen_servicios);
         imagen.setClickable(true);
-        Intent intent = new Intent(this, MisAlarmas1.class);
+        Intent intent = new Intent(this, Servicios.class);
         startActivity(intent);
+        finish();
     }
 
     public void irPerfil(View view) {
@@ -117,6 +119,7 @@ public class Mi_Cuenta extends AppCompatActivity {
         imagen.setClickable(true);
         Intent intent = new Intent(this, Mi_Cuenta.class);
         startActivity(intent);
+        finish();
     }
 
     public void irContacto(View view) {
@@ -124,6 +127,7 @@ public class Mi_Cuenta extends AppCompatActivity {
         imagen.setClickable(true);
         Intent intent = new Intent(this, ContactoActivity.class);
         startActivity(intent);
+        finish();
     }
 
     @Override
