+++ app/src/main/java/com/dpozzo68/analizarapp/MisAlarmas1.java
@@ -71,41 +71,45 @@ public class MisAlarmas1 extends AppCompatActivity{
     public void irAlarmasConfiguracion(){
         Intent intent = new Intent(this, AlarmasConfiguracionActivity.class);
         startActivity(intent);
+        finish();
     }
 
-    public void irConsumos(View view) {
-        ImageView imagen = findViewById(R.id.imagen_home);
-        imagen.setClickable(true);
-        Intent intent = new Intent(this, MisConsumos.class);
-        startActivity(intent);
-    }
-
-    public void irAlarmas(View view) {
-        ImageView imagen = findViewById(R.id.imagen_alerta);
-        imagen.setClickable(true);
-        Intent intent = new Intent(this, MisAlarmas1.class);
-        startActivity(intent);
-    }
-
-    public void irPerfil(View view) {
-        ImageView imagen = findViewById(R.id.imagen_perfil);
-        imagen.setClickable(true);
-        Intent intent = new Intent(this, Mi_Cuenta.class);
-        startActivity(intent);
-    }
-
-    public void irContacto(View view) {
-        ImageView imagen = findViewById(R.id.imagen_contacto);
-        imagen.setClickable(true);
-        Intent intent = new Intent(this, ContactoActivity.class);
-        startActivity(intent);
-    }
+//    public void irConsumos(View view) {
+//        ImageView imagen = findViewById(R.id.imagen_home);
+//        imagen.setClickable(true);
+//        Intent intent = new Intent(this, MisConsumos.class);
+//        startActivity(intent);
+//        finish();
+//    }
+//
+//    public void irServicios(View view) {
+//        ImageView imagen = findViewById(R.id.imagen_servicios);
+//        imagen.setClickable(true);
+//        Intent intent = new Intent(this, Servicios.class);
+//        startActivity(intent);
+//        finish();
+//    }
+//
+//    public void irPerfil(View view) {
+//        ImageView imagen = findViewById(R.id.imagen_perfil);
+//        imagen.setClickable(true);
+//        Intent intent = new Intent(this, Mi_Cuenta.class);
+//        startActivity(intent);
+//        finish();
+//    }
+//
+//    public void irContacto(View view) {
+//        ImageView imagen = findViewById(R.id.imagen_contacto);
+//        imagen.setClickable(true);
+//        Intent intent = new Intent(this, ContactoActivity.class);
+//        startActivity(intent);
+//        finish();
+//    }
 
     @Override
     public void onBackPressed() {
-        Intent intent = new Intent(this, MisConsumos.class);
+        Intent intent = new Intent(this, Servicios.class);
         startActivity(intent);
-        finish();
     }
